requirements.txt
# 🤖 X自動反応ツール - Python依存関係
# 運営者ブラインド設計・プライバシー重視

# ==============================================================================
# 🚀 Webフレームワーク・API
# ==============================================================================

# FastAPI - 高性能WebAPIフレームワーク
fastapi==0.104.1
uvicorn[standard]==0.24.0

# ASGI・WSGI サーバー（本番環境用）
gunicorn==21.2.0

# ==============================================================================
# 🗄️ データベース・ORM
# ==============================================================================

# SQLAlchemy - ORM
sqlalchemy==2.0.23
alembic==1.12.1

# 非同期サポート
asyncpg==0.29.0        # PostgreSQL（Render用）
aiosqlite==0.19.0      # SQLite（開発用）

# データベースドライバー
psycopg2-binary==2.9.9  # PostgreSQL（本番環境用）

# ==============================================================================
# 🔐 暗号化・セキュリティ
# ==============================================================================

# 暗号化ライブラリ
cryptography==41.0.7
pycryptodome==3.19.0

# パスワードハッシュ化
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# JWT認証
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# ==============================================================================
# 🤖 AI・機械学習
# ==============================================================================

# Groq AI クライアント
groq==0.4.1

# HTTP クライアント（AI API用）
httpx==0.25.2
aiohttp==3.9.1

# ==============================================================================
# 🐦 X (Twitter) API
# ==============================================================================

# Tweepy - X API クライアント
tweepy==4.14.0

# OAuth認証
requests-oauthlib==1.3.1
oauthlib==3.2.2

# ==============================================================================
# 🌐 HTTP・ネットワーク
# ==============================================================================

# HTTP リクエスト
requests==2.31.0
urllib3==2.1.0

# CORS対応
fastapi-cors==0.0.6

# ==============================================================================
# ⏰ 日時・スケジューリング
# ==============================================================================

# 日時処理
python-dateutil==2.8.2
pytz==2023.3.post1

# スケジューリング
schedule==1.2.0
APScheduler==3.10.4

# ==============================================================================
# 📊 データ処理
# ==============================================================================

# JSON・データ処理
pydantic==2.5.1
pydantic-settings==2.1.0

# データ検証
email-validator==2.1.0.post1

# ==============================================================================
# 📝 ログ・監視
# ==============================================================================

# 構造化ログ
structlog==23.2.0
python-json-logger==2.0.7

# 監視・メトリクス
prometheus-client==0.19.0

# ==============================================================================
# 🔧 設定・環境管理
# ==============================================================================

# 環境変数管理
python-dotenv==1.0.0

# 設定管理
pyyaml==6.0.1
toml==0.10.2

# ==============================================================================
# 🧪 テスト・開発
# ==============================================================================

# テストフレームワーク
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# モック・テストデータ
httpx==0.25.2  # テスト用HTTPクライアント
faker==20.1.0  # テストデータ生成

# ==============================================================================
# 🚀 本番環境・デプロイメント
# ==============================================================================

# プロセス管理
supervisor==4.2.5

# メモリ・パフォーマンス監視
memory-profiler==0.61.0
psutil==5.9.6

# ==============================================================================
# 🔄 非同期・並行処理
# ==============================================================================

# 非同期処理
asyncio==3.4.3
aiofiles==23.2.1

# タスクキュー
celery==5.3.4
redis==5.0.1

# ==============================================================================
# 📡 外部サービス連携
# ==============================================================================

# メール送信
emails==0.6

# Slack通知（オプション）
slack-sdk==3.26.1

# ==============================================================================
# 🛡️ セキュリティ・バリデーション
# ==============================================================================

# 入力値検証
validators==0.22.0
bleach==6.1.0

# セキュリティヘッダー
secure==0.3.0

# ==============================================================================
# 📦 ユーティリティ
# ==============================================================================

# 型ヒント
typing-extensions==4.8.0

# 進捗バー
tqdm==4.66.1

# カラー出力
colorama==0.4.6
rich==13.7.0

# ==============================================================================
# 💰 本番環境最適化
# ==============================================================================

# メモリ効率化
lru-dict==1.3.0

# キャッシュ
cachetools==5.3.2

# 圧縮
gzip==1.0

# ==============================================================================
# 🌐 Render 特化
# ==============================================================================

# Render 環境での最適化
whitenoise==6.6.0      # 静的ファイル配信
dj-database-url==2.1.0 # データベースURL解析

# ==============================================================================
# 📋 バージョン固定理由
# ==============================================================================

# fastapi==0.104.1      # 安定版、セキュリティアップデート対応
# sqlalchemy==2.0.23    # 最新の非同期サポート
# cryptography==41.0.7  # セキュリティ修正済み
# tweepy==4.14.0        # X API v2 対応
# groq==0.4.1           # 最新AI機能
# pydantic==2.5.1       # パフォーマンス向上版

# ==============================================================================
# 🔄 定期更新推奨パッケージ
# ==============================================================================

# セキュリティ関連は定期的にアップデート:
# - cryptography
# - requests
# - urllib3
# - pyjwt
# - sqlalchemy

# ==============================================================================
# 💡 使用方法
# ==============================================================================

# インストール:
# pip install -r requirements.txt

# 開発環境:
# pip install -r requirements.txt
# pip install -e .

# 本番環境（Render）:
# pip install --no-dev -r requirements.txt

# 依存関係確認:
# pip list --outdated
# pip check

# セキュリティ監査:
# pip audit
# safety check

# ==============================================================================
# 🔐 セキュリティノート
# ==============================================================================

# 1. 定期的なセキュリティアップデート必須
# 2. 脆弱性スキャンの実行
# 3. 不要なパッケージの削除
# 4. バージョン固定によるreproducible builds
# 5. 最小権限の原則（必要最小限のパッケージのみ）